Refactored polynomial operations with cythonized implementations into `objects.opcalc` subpackage
