+++ packages/pygsti/objects/polynomial.py
@@ -203,7 +203,7 @@ class FASTPolynomial(object):
         Generate a compact form of this polynomial designed for fast evaluation.
 
         The resulting "tapes" can be evaluated using
-        :function:`bulk_eval_compact_polys`.
+        :function:`opcalc.bulk_eval_compact_polys`.
 
         Parameters
         ----------
@@ -612,7 +612,7 @@ class SLOWPolynomial(dict):  # REMOVE THIS CLASS (just for reference)
         Generate a compact form of this polynomial designed for fast evaluation.
 
         The resulting "tapes" can be evaluated using
-        :function:`bulk_eval_compact_polys`.
+        :function:`opcalc.bulk_eval_compact_polys`.
 
         Parameters
         ----------
@@ -955,64 +955,6 @@ class SLOWPolynomial(dict):  # REMOVE THIS CLASS (just for reference)
         return replib.PolyRep(int_coeffs, max_num_vars, vindices_per_int)
 
 
-def bulk_eval_compact_polys(vtape, ctape, paramvec, dest_shape, dtype="auto"):
-    """
-    Evaluate many compact polynomial forms at a given set of variable values.
-
-    Parameters
-    ----------
-    vtape, ctape : numpy.ndarray
-        Specifies "variable" and "coefficient" 1D numpy arrays to evaluate.
-        These "tapes" can be generated by concatenating the tapes of individual
-        complact-polynomial tuples returned by :method:`Polynomial.compact`.
-
-    paramvec : array-like
-        An object that can be indexed so that `paramvec[i]` gives the
-        numerical value to substitute for i-th polynomial variable (x_i).
-
-    dest_shape : tuple
-        The shape of the final array of evaluated polynomials.  The resulting
-        1D array of evaluated polynomials is reshaped accordingly.
-
-    dtype : {"auto", "real", "complex}
-        The type of the coefficient array that is returned.
-
-    Returns
-    -------
-    numpy.ndarray
-        An array of the same type as the coefficient tape or with the type
-        given by `dtype`, and with shape given by `dest_shape`.
-    """
-    if dtype == "auto":
-        result = _np.empty(dest_shape, ctape.dtype)  # auto-determine type?
-    elif dtype == "complex":
-        result = _np.empty(dest_shape, complex)
-    elif dtype == "real":
-        result = _np.empty(dest_shape, 'd')
-    else:
-        raise ValueError("Invalid dtype: %s" % dtype)
-
-    res = result.flat  # for 1D access
-
-    c = 0; i = 0; r = 0
-    while i < vtape.size:
-        poly_val = 0
-        nTerms = vtape[i]; i += 1
-        #print("POLY w/%d terms (i=%d)" % (nTerms,i))
-        for m in range(nTerms):
-            nVars = vtape[i]; i += 1  # number of variable indices in this term
-            a = ctape[c]; c += 1
-            #print("  TERM%d: %d vars, coeff=%s" % (m,nVars,str(a)))
-            for k in range(nVars):
-                a *= paramvec[vtape[i]]; i += 1
-            poly_val += a
-            #print("  -> added %s to poly_val = %s" % (str(a),str(poly_val))," i=%d, vsize=%d" % (i,vtape.size))
-        res[r] = poly_val; r += 1
-    assert(c == ctape.size), "Coeff Tape length error: %d != %d !" % (c, ctape.size)
-    assert(r == result.size), "Result/Tape size mismatch: only %d result entries filled!" % r
-    return result
-
-
 def bulk_load_compact_polys(vtape, ctape, keep_compact=False, max_num_vars=100):
     """
     Create a list of Polynomial objects from a "tape" of their compact versions.
@@ -1059,97 +1001,3 @@ def bulk_load_compact_polys(vtape, ctape, keep_compact=False, max_num_vars=100):
                 poly_coeffs[tuple(vtape[i:i + nVars])] = a; i += nVars
             result.append(Polynomial(poly_coeffs, max_num_vars))
     return result
-
-
-def compact_deriv(vtape, ctape, wrtParams):
-    """
-    Take the derivative of one or more compact Polynomials with respect
-    to one or more variables/parameters.
-
-    Parameters
-    ----------
-    vtape, ctape : numpy.ndarray
-        Specifies "variable" and "coefficient" 1D numpy arrays to differentiate.
-        These "tapes" can be generated by concatenating the tapes of individual
-        complact-polynomial tuples returned by :method:`Polynomial.compact`.
-
-    wrtParams : list
-        The variable indices to differentiate with respect to.  They
-        must be sorted in ascending order. E.g. "[0,3]" means separatey
-        differentiate w.r.t x_0 and x_3 (concatenated first by wrtParam
-        then by poly).
-
-    Returns
-    -------
-    vtape, ctape : numpy.ndarray
-    """
-    result_vtape = []
-    result_ctape = []
-    wrt = sorted(wrtParams)
-    assert(wrt == list(wrtParams)), "`wrtParams` (%s) must be in ascending order!" % wrtParams
-    #print("TAPE SIZE = ",vtape.size)
-
-    c = 0; i = 0
-    while i < vtape.size:
-        j = i  # increment j instead of i for this poly
-        nTerms = vtape[j]; j += 1
-        dctapes = [list() for x in range(len(wrt))]
-        dvtapes = [list() for x in range(len(wrt))]
-        dnterms = [0] * len(wrt)
-        #print("POLY w/%d terms (i=%d)" % (nTerms,i))
-        for m in range(nTerms):
-            coeff = ctape[c]; c += 1
-            nVars = vtape[j]; j += 1  # number of variable indices in this term
-
-            #print("  TERM%d: %d vars, coeff=%s" % (m,nVars,str(coeff)))
-            cur_iWrt = 0
-            j0 = j  # the vtape index where the current term starts
-
-            #Loop to get counts of each variable index that is also in `wrt`.
-            # Once we've passed an element of `wrt` process it, since there can't
-            # see it any more (the var indices are sorted).
-            while j < j0 + nVars:  # loop over variable indices for this term
-                # can't be while True above in case nVars == 0 (then vtape[j] isn't valid)
-
-                #find an iVar that is also in wrt.
-                # - increment the cur_iWrt or j as needed
-                while cur_iWrt < len(wrt) and vtape[j] > wrt[cur_iWrt]:  # condition to increment cur_iWrt
-                    cur_iWrt += 1  # so wrt[cur_iWrt] >= vtape[j]
-                if cur_iWrt == len(wrt): break  # no more possible iVars we're interested in;
-                # we're done with all wrt elements
-                # - at this point we know wrt[cur_iWrt] is valid and wrt[cur_iWrt] >= tape[j]
-                while j < j0 + nVars and vtape[j] < wrt[cur_iWrt]:
-                    j += 1  # so vtape[j] >= wrt[cur_iWrt]
-                if j == j0 + nVars: break  # no more iVars - we're done
-
-                #print(" check j=%d, val=%d, wrt=%d, cur_iWrt=%d" % (j,vtape[j],wrt[cur_iWrt],cur_iWrt))
-                if vtape[j] == wrt[cur_iWrt]:
-                    #Yay! a value we're looking for is present in the vtape.
-                    # Figure out how many there are (easy since vtape is sorted
-                    # and we'll always stop on the first one)
-                    cnt = 0
-                    while j < j0 + nVars and vtape[j] == wrt[cur_iWrt]:
-                        cnt += 1; j += 1
-                    #Process cur_iWrt: add a term to tape for cur_iWrt
-                    dvars = list(vtape[j0:j - 1]) + list(vtape[j:j0 + nVars])  # removes last wrt[cur_iWrt] var
-                    dctapes[cur_iWrt].append(coeff * cnt)
-                    dvtapes[cur_iWrt].extend([nVars - 1] + dvars)
-                    dnterms[cur_iWrt] += 1
-                    # print(" wrt=%d found cnt=%d: adding deriv term coeff=%f vars=%s" \
-                    #       % (wrt[cur_iWrt], cnt, coeff*cnt, [nVars-1] + dvars))
-
-                    cur_iWrt += 1  # processed this wrt param - move to next one
-
-            #Now term has been processed, adding derivative terms to the dctapes and dvtapes "tape-lists"
-            # We continue processing terms, adding to these tape lists, until all the terms of the
-            # current poly are processed.  Then we can concatenate the tapes for each wrtParams element.
-            # Move to next term; j may not have been incremented if we exited b/c of cur_iWrt reaching end
-            j = j0 + nVars
-
-        #Now all terms are processed - concatenate tapes for wrtParams and add to resulting tape.
-        for nTerms, dvtape, dctape in zip(dnterms, dvtapes, dctapes):
-            result_vtape.extend([nTerms] + dvtape)
-            result_ctape.extend(dctape)
-        i = j  # update location in vtape after processing poly - actually could just use i instead of j it seems??
-
-    return _np.array(result_vtape, _np.int64), _np.array(result_ctape, complex)
